Implement the disk image change mechanism.
If GRAPH is pressed then CAPS lits and waits for a number or a
letter (A-Z) and then mounts the Nth file corresponding to the
key pressed (1-9, A=10, B=11, etc) in the same directory as the
currently mounted file. Pressing GRAPH again cancels the operation.

Additionally, the corresponding key can be pressed directly on
disk access, without pressing GRAPH.