Fix: DSKCHG crashing in disk emulation mode in some games (#11)
* Fix: DSKCHG crashing in disk emulation mode in some games

The exact cause of the crash is unknown, but it's at some point
in MAYBE_CHANGE_DSK. So no we don't call it and always return
a disk change status of "unknown".

* Fix: "Enable/Disable CAPS lit on disk access" display in boot menu

The text wasn't appearing in the config menu for directories,
but it was still possible to press 8 and it had the expected effect.